Started Migration of MVVM from MLib2 to MLib3.
There will be a massive cleanup in the library to address performance issues and to focus on best practices (without having too much choices to do otherwise (like before)).
The very core of the lib will change and some functions like "CreateCommandBindings" will be removed to enforce cleaner code (and code analysis) in the future - and to get rid of some slow reflection stuff.